requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2022.9.24
chardet==4.0.0
charset-normalizer==2.1.1
Django==4.1.2
exceptiongroup==1.0.0rc9
googlesearch-python==1.1.0
h11==0.14.0
idna==2.10
lxml==4.9.1
numpy==1.23.4
outcome==1.2.0
PySocks==1.7.1
python-dotenv==0.21.0
requests==2.25.1
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.12
w3lib==2.0.1
webdriver-manager==3.8.3
wsproto==1.2.0